fiddle shell format for the Nth time
Galaxy makes several complaints about my line length.  It has a point,
the FPRINT= line should not have been so long.  Break out $VER; that
makes it more consistent with the other shell blocks.

At the same time, I think at least error messages will be clearer, if we
use the new-line preserving string mode.  And it avoids the strangeness
of having to put ";" at the end of the line!  I've used that for other
roles, so let's use it here as well.