update_master_keys fix for current Ansible version
fetch module has stopped ignoring check mode, now it skips as expected
(e.g. compare command module).  Change our task so it always runs, as it
did before.

The alternative would be to fail if ansible_check_mode is true.
update_master_keys is a hack for maintainers only,
so this maintainer chooses the arguably evil behaviour which is most
convenient for him.